+++ src/users/dto/fetch-user.dto.ts
@@ -1,20 +1,24 @@
-import { IsOptional, IsString } from 'class-validator';
+import { IsEnum, IsOptional, IsString } from 'class-validator';
 import { Mbti, Position } from '../users.enum';
 
-export class FetchUsersDto {
+export class FetchUserDto {
   @IsOptional()
   id: number;
 
   @IsOptional()
+  @IsString()
   name: string;
 
   @IsOptional()
+  @IsEnum(Position)
   position: Position;
 
   @IsOptional()
+  @IsEnum(Mbti)
   mbti: Mbti;
 }
-export class FetchUsersParams {
+
+export class FetchUserParam {
   id?: number;
 
   name?: string;
